IEI-207902 Improved the duplicate UUID reference check in caching decorator
Also added debug and trace level logging to aid in future troubleshooting
